stop using SPI to enable/disable drv8711
I don't think we should be enabling/disabling the drv8711 via SPI commands this way. Connecting the !sleep pin on the driver IC to the en signal for the  chosen driver channel is, imhop, the right way to control the enabling/disabling of the driver. Sending these SPI commands is actually breaking things in my setup in practice: re-enabling after disabling is not working. I assume this problem I'm having is indicative of some other issue in my setup, but I can't find it (maybe there's some interaction between the enable bit and the sleep pin being changed at essentially the same time?). With the change here, everything works fine. Using the !sleep pin properly powers and unpowers the driver and everything works as expected.