drv8711, enable/disable should be by hardware line
I don't think we should be enabling/disabling the drv8711 via SPI commands. It seems cleaner, more reliable and in line with how the built in drivers work to use a hardware enable line tied to the !sleep pin on the drv8711. Sending these SPI commands is actually breaking things in my setup in practice: re-enabling after disabling is not working. Using the change here, everything works fine (provided the driver channel's enable signal is properly setup in the config file and tied to the drv8711's !sleep pin).